🗺️ Atlas: break db <-> query circular dependency
* Extracted `execute_query` to `GraphView` trait in `src/query/traits.rs`.
* Implemented `execute_query` for `AletheiaDB` in `src/db/graph_view.rs`.
* Changed `QueryBuilder::execute` to accept `&dyn GraphView` rather than concrete `AletheiaDB`.
* Moved all tests in `hybrid.rs` and `semantic_pathfinding.rs` that rely on concrete `AletheiaDB` instantiation to `src/db/tests.rs` to break the cycle completely.
* Tested with `cargo check` and `cargo test`.
